Added a new idea -> refPoint to the gameState
Some objects need a point of reference. The new variable refPoint will
be that point of reference. In this game will be the position of the
hero but in another game we can take another point of reference. For
example, that point of reference is used in the default cameraand in the
notification of the graphic engine